requirement.txt
streamlit
pandas
numpy
scikit-learn
opencv-python
Pillow